pyproject.toml
[tool.poetry]
name = "pyfuta"
version = "0.1.0"
description = "A low code analysis platform based on python."
authors = ["Usagi no Niku <chenbohan911@163.com>"]
license = "GPL-3.0"
readme = "README.md"
packages = [
    { include = "pyfuta" }
]

[tool.poetry.dependencies]
python = "3.8.6"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
python-multipart = "^0.0.9"
uvicorn = "^0.30.0"
aiohttp = "^3.9.5"
orjson = "^3.10.3"
sse-starlette = "^2.1.0"
sqlmodel = "^0.0.18"
tzdata = "^2024.1"
aiosqlite = "^0.20.0"
openpyxl = "^3.1.5"
pyinstaller = "^6.10.0"

[tool.poetry.scripts]
build = "pyfuta.pyinstaller:install"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
